Move matrix class to quality_estimator
