+++ src/translator/CMakeLists.txt
@@ -14,7 +14,6 @@ add_library(bergamot-translator STATIC
     annotation.cpp
     service.cpp
     threadsafe_batcher.cpp
-    matrix.cpp
 )
 if (USE_WASM_COMPATIBLE_SOURCE)
   # Using wasm compatible sources should include this compile definition;
